requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1
alpaca-py==0.33.1
httpx==0.28.1
sqlalchemy==2.0.36
aiosqlite==0.20.0
python-dotenv==1.0.1
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-httpx==0.35.0
loguru==0.7.3
pytz==2024.2
psycopg2-binary==2.9.10
asyncpg==0.30.0
